Add url pattern for update_certificate
Evidently the tool already supports updating a certificate, but for some reason the url pattern was never implemented. This feature will be useful if a student takes an exam and passes on the first try with a minimum grade of 85%, but decides to take the exam again to get a better grade. The current implementation logs the first exam grade in Accredible. See https://github.com/gymnasium/tracker/issues/44 for additional details